requirements.txt
beautifulsoup4==4.12.3
chromadb==0.4.24
streamlit==1.33.0
streamlit-float==0.3.2
langchain==0.1.16
langchain_community==0.0.32
langchain_openai==0.1.3
langchain_core==0.1.42
langchain_google_genai==0.0.6
python-dotenv==1.0.0